ava-api-server/ava/requirements.txt
beautifulsoup4==4.12.2
exchangelib==5.2.0
Flask==3.0.2
flask_cors==4.0.0
flask_session==0.6.0
flask-redis==0.4.0
Flask-SQLAlchemy==3.1.1
gevent==24.2.1
kor==3.0.0
langchain==0.3.22
langchain-community==0.3.20
langchain-core==0.3.49
langchain-openai==0.3.11
openai==1.68.2
psycopg2==2.9.9
psycopg2_binary==2.9.9
psycopg2-pool==1.1
cors==1.0.1
pyDes==2.0.1
pydub==0.25.1
python-dotenv==1.0.0
PyYAML==6.0.1
redis==5.0.0
Requests==2.31.0
retrying==1.3.4
streamlit==1.44.1
tabulate==0.9.0
gunicorn==21.2.0
urllib3==2.2.1
PyJWT==2.8.0
pypdf==4.1.0
pytz==2023.3
anthropic==0.21.3
ragas==0.1.7
pgvector==0.2.5
pytest==8.2.1
pytest-order==1.2.1
prometheus_flask_exporter==0.23.1
loki-logger-handler==0.1.4
aiopg==1.4.0
google-genai==1.2.0
mistralai==1.6.0
BabelDOC==0.2.30
jieba==0.42.1
torch==2.7.0
torchvision==0.22.0
pillow==11.2.1
doclayout_yolo==0.0.3
./metadata_extractor-0.2.5-py3-none-any.whl